% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportToCSV.R
\name{exportResultsToCSV}
\alias{exportResultsToCSV}
\title{The main CatalogueExport analyses (for v5.x)}
\usage{
exportResultsToCSV(
  connectionDetails,
  resultsDatabaseSchema,
  analysisIds = c(),
  smallCellCount = 5,
  exportFolder = "./output"
)
}
\arguments{
\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}

\item{resultsDatabaseSchema}{Fully qualified name of database schema that we can write final results to. Default is cdmDatabaseSchema. 
On SQL Server, this should specifiy both the database and the schema, so for example, on SQL Server, 'cdm_results.dbo'.}

\item{analysisIds}{A vector containing the set of CatalogueExport analysisIds for which results will be generated.}

\item{smallCellCount}{To avoid patient identifiability, cells with small counts (<= smallCellCount) are deleted. Set to NULL if you don't want any deletions.}

\item{exportFolder}{Folder to export the results to.}
}
\description{
\code{exportResults} exports the results to a csv file for upload to the Catalogue
}
