% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helper.R
\name{printAnalysesSql}
\alias{printAnalysesSql}
\title{Print the sql of an analysis}
\usage{
printAnalysesSql(analysisId, connectionDetails)
}
\arguments{
\item{analysisId}{An analysisId for which the sql will be printed.}

\item{connectionDetails}{An R object of type \code{connectionDetails} created using the function \code{createConnectionDetails} in the \code{DatabaseConnector} package.}
}
\value{
None
}
\description{
Print the sql of an analysis
}
\details{
Print the parameterized SQL that is run for an analysisId.
}
